Исправление ошибок 2: В task и его наследниках: - исправление проверок на id - добавление клонирующих конструкторов - исправление методов обновление информации
