+++ src/Tasks/Subtask.java
@@ -8,6 +8,12 @@ public class Subtask extends Task {
         this.epicId = epicId;
     }
 
+    // клонирующий конструктор
+    public Subtask(Subtask subtask) {
+        super(subtask);
+        this.epicId = subtask.getEpicId();
+    }
+
 
     public int getEpicId() {
         return epicId;
